backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
pymongo==4.5.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
cryptography==41.0.4
python-dotenv==1.0.0
requests==2.31.0
pyjwt==2.8.0
geopy==2.4.0
pyproj==3.6.1
